Idiot-proof against destructive symlinking

+++ .install.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
 
+if [ $PWD == $HOME ]; then
+  echo "Run this from your dotfiles repo, not your home directory!"
+  exit 0
+fi
+
 ln -sF $PWD/.* $HOME
 unlink $HOME/.git
 unlink $HOME/.gitignore
